+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.6.0] Q3 2022
- - renamed crate from demo-portal-module to remote-features-module
+ - renamed crate from demo-portal-module to icsdm-device-service
 
 ## [0.5.11] Q3 2022
  - added direct method "reboot"